Ready for testing.

+++ cmd/elevator/main.go
@@ -1,65 +1,107 @@
 package main
 
 import (
-	"fmt"
+	"flag"
+	"log"
 	"net"
 	"os"
+	"os/signal"
+	"syscall"
 	"time"
-	//"os/signal"
 
 	"elevator-project/pkg/config"
 	"elevator-project/pkg/elev"
 	"elevator-project/pkg/network"
 )
 
+var (
+        noWatchdog = flag.Bool("nowatchdog", false,
+                "Set this to run without a watchdog process.")
+)
+
+// Logging
+var debug *log.Logger
+var errorlog *log.Logger
+
+func init() {
+	debug = log.New(os.Stdout, "", 0)
+	errorlog = log.New(os.Stderr, "\x1b[31mERROR\x1b[m: ", 0)
+}
+
+
 type ServiceMode int
 
 const (
 	Online ServiceMode = 0x1
 	Local  ServiceMode = 0x2
-	Broken ServiceMode = 0x3
 )
 
+// Stores backups for all connected elevators.
 type BackupHandler struct {
 	backups map[network.Addr]*backupData
 	addr    network.Addr
+
+	// True when the latest backup has been synchronized with the other elevators.
+	// When service mode is online and synced is false, the synchronization is not
+	// complete. 
 	synced  bool
+
+	// An empty struct is sent on this channel when the latest backup
+	// and the current elevator state is different and synced is true.
+	// This means that reading from this channel blocks while synchronizing.
+	invalid chan struct{}
 }
 
+// Copy the elevator state into a backupData struct, and store in backup database.
 func (b *BackupHandler) create(e *Elevator) *backupData {
 	var bd = &backupData{
-		elevator: b.addr,
-		created:  time.Now(),
-		requests: e.requests,
-		dest:     e.dest,
+		elevator:  b.addr,
+		created:   time.Now(),
+		floor:     e.floor,
+		direction: e.direction,
+		requests:  e.requestsBuffer,
+		dest:      e.destBuffer,
 	}
 	b.backups[b.addr] = bd
 	return bd
 }
 
+// Get the latest backup of this elevator.
 func (b *BackupHandler) get() *backupData {
 	return b.backups[b.addr]
 }
 
+// Store backupData struct in database.
 func (b *BackupHandler) update(bd *backupData) {
 	b.backups[bd.elevator] = bd
 }
 
+// Check if current elevator state differs from latest backup. Runs in a goroutine.
 func (b *BackupHandler) changed(e *Elevator) bool {
-	backup := b.backups[b.addr]
-	return !(e.requests == backup.requests && e.dest == backup.dest)
+	for {
+		backup := b.backups[b.addr]
+		if !(e.requestsBuffer == backup.requests && e.destBuffer == backup.dest) && b.synced {
+			b.invalid <- struct{}{}
+		}
+	}
 }
 
+// Handles communication with the watchdog process over Unix domain sockets.
 type WatchdogHandler struct {
 	conn     *net.UnixConn
 	watchdog *net.UnixAddr
 	addr     *net.UnixAddr
-	timer    Timer
+	timer    *time.Timer
 }
 
-const watchdogResendInterval = 100 * time.Millisecond
+const watchdogResendInterval = 150 * time.Millisecond
 
-func (wd *WatchdogHandler) start() ([]byte, error) {
+// Connects to watchdog process and loads inital backup.
+func (wd *WatchdogHandler) start() (*backupData, error) {
+	if *noWatchdog {
+		return &backupData{}, nil
+	}
+	
 	// unlink socket
 	os.Remove(wd.addr.Name)
 
@@ -81,12 +123,19 @@ func (wd *WatchdogHandler) start() ([]byte, error) {
 		return nil, err
 	}
 
-	return buf[:n], nil
+	bd := &backupData{}
+	unpackData(buf[:n], bd)
+	
+	return bd, nil
 }
 
+// Send backupData to watchdog process.
 func (wd *WatchdogHandler) writeBackup(bd *backupData) error {
-	data, _ := bd.MarshalBinary()
+	if *noWatchdog {
+		return nil
+	}
 
+	data, _ := bd.MarshalBinary()
 	_, err := wd.conn.WriteToUnix(data, wd.watchdog)
 	if err != nil {
 		return err
@@ -95,177 +144,299 @@ func (wd *WatchdogHandler) writeBackup(bd *backupData) error {
 }
 
 func main() {
+	flag.Parse()
+
+	var mode ServiceMode
+
+	// Load configuration file.
 	conf, _ := config.LoadFile("./config")
 	elev.LoadConfig(conf)
 	network.LoadConfig(conf)
 
+	// Initialize WatchdogHandler and load elevator backup.
 	watchdog := &WatchdogHandler{
 		watchdog: &net.UnixAddr{conf["watchdog.socket"], "unixgram"},
 		addr:     &net.UnixAddr{conf["watchdog.elev_socket"], "unixgram"},
+		timer:    time.NewTimer(watchdogResendInterval),
 	}
-	watchdog.start()
+	wdbackup, _ := watchdog.start()
 
+	// Initialize elavator hardware.
 	err := elev.Init()
 	if err != nil {
-		fmt.Println(err)
+		debug.Println(err)
 		os.Exit(1)
 	}
 
+	// Create objects.
 	node := network.NewNode()
-	node.Start()
-
 	panel := NewPanel()
-	panel.Start()
-
 	elevator := NewElevator(panel)
-	elevator.Start()
 
+	// Load the backup from the watchdog process. This does nothing if
+	// wdbackup has only nil-values.
+	elevator.LoadBackup(wdbackup)
+	panel.LoadBackup(wdbackup)
+
+	// Setup channels.
 	msgsFromOther := make(chan *network.Message)
 	msgsFromThis := make(chan *network.Message)
+	deadNode := make(chan network.Addr)
+
+	// Setup signal handler
+	interrupt := make(chan os.Signal, 1)
+	signal.Notify(interrupt, syscall.SIGINT)
+
+	// Start all threads. NOTE!! The order matters.
+	node.Start()
+	panel.Start()
+	elevator.Start()
+
+	// Start Goroutines that put all incomming messages onto channels.
 	go receiveMsgs(node, msgsFromOther)
 	go receiveMyMsgs(node, msgsFromThis)
+	go getDeadNode(node, deadNode)
 
-	var backup = &BackupHandler{make(map[network.Addr]*backupData), node.Addr(), false}
+	// Initialize BackupHandler and store an initial backup.
+	var backup = &BackupHandler{
+		backups: make(map[network.Addr]*backupData),
+		addr:    node.Addr(),
+		synced:  true,
+		invalid: make(chan struct{}),
+	}
 	backup.create(elevator)
+	go backup.changed(elevator)
 
-	var mode ServiceMode = Local
-	reqch := panel.Requests
+	// Start elevator in local mode.
+	mode = Local
+
+	// Channel containing all unassigned requests.
+	unassigned := make(chan Request, 32)
+
+	// When an unassigned request is being processed, reqch is set to nil. This means
+	// we process only one request at a time.
+	reqch := unassigned
+
+	// The current unassigned request being processed.
+	var req Request
 
 	for {
-		// TODO: check node.IsConnected()
-		
-		if watchdog.timer.HasTimedOut() {
+
+		if node.IsConnected() {
+			mode = Online
+		} else {
+			mode = Local
+
+			// Check if the elevator was disconnected in the middle of a transaction.
+			if reqch == nil {
+				unassigned <- req
+			}
+		}
+
+		select {
+
+		case <-watchdog.timer.C:
 			watchdog.writeBackup(backup.get())
 			watchdog.timer.Reset(watchdogResendInterval)
-		}
-		
-		if mode == Online {
 
-			if backup.synced && backup.changed(elevator) {
-				bd := backup.create(elevator)
+		case <-backup.invalid:
+			bd := backup.create(elevator)
+			watchdog.writeBackup(bd)
+
+			debug.Printf("Backup message: \n\t%v\n", bd)
+
+			if mode == Online && backup.synced {
 				backup.synced = false
 				sendData(node, BACKUP, bd)
+				debug.Printf("Sent backup message: \n\t%v\n", bd)
 			}
 
-			select {
-			case req := <-reqch:
-				sendData(node, PANEL, &panelData{
-					floor:  req.Floor,
-					button: btnFromDir(req.Direction),
-					on:     true,
-				})
+		case r := <-panel.Requests:
+			unassigned <- r
 
-				fmt.Printf("PANEL message sent\n")
-
-				sendData(node, COST, &costData{
+		case req = <-reqch:
+			if mode == Online {
+				var cd = costData{
 					elevator: node.Addr(),
 					req:      req,
 					cost:     elevator.SimulateCost(req),
-				})
-
-				fmt.Printf("COST message sent\n")
+				}
+				sendData(node, COST, &cd)
+				debug.Printf("Sent cost message: \n\t%v\n", cd)
 
 				reqch = nil // handle only Request at a time.
-
-			case msg := <-msgsFromOther:
-				switch msg.Type {
-				case PANEL:
-					var pd panelData
-					unpackData(msg.Data, &pd)
-					panel.SetLamp(pd.button, pd.floor, pd.on)
-					fmt.Printf("PANEL message received and forwarded\n")
-
-				case COST:
-					var cd costData
-					unpackData(msg.Data, &cd)
-					cost := elevator.SimulateCost(cd.req)
-					if cost < cd.cost {
-						cd.elevator = node.Addr()
-						cd.cost = cost
-					}
-					packData(msg.Data, &cd)
-
-					fmt.Printf("COST message received and forwarded\n")
-
-				case ASSIGN:
-					var ad assignData
-					unpackData(msg.Data, &ad)
-					if ad.elevator == node.Addr() {
-						elevator.AddRequest(ad.req)
-						ad.taken = true
-						fmt.Printf("ASSIGN message taken\n")
-					}
-					packData(msg.Data, &ad)
-
-					fmt.Printf("ASSIGN message received and forwarded\n")
-
-				case BACKUP:
-					bd := &backupData{}
-					unpackData(msg.Data, bd)
-					backup.update(bd)
-
-				}
-				node.ForwardMessage(msg)
-
-			case msg := <-msgsFromThis:
-				switch msg.Type {
-				case COST:
-					var cd costData
-					unpackData(msg.Data, &cd)
-					sendData(node, ASSIGN, &assignData{
-						elevator: cd.elevator,
-						req:      cd.req,
-					})
-					fmt.Printf("COST message received from myself\n")
-
-				case ASSIGN:
-					var ad assignData
-					unpackData(msg.Data, &ad)
-					reqch = panel.Requests
-					fmt.Printf("ASSIGN message received from myself with taken = %v\n", ad.taken)
-					if !ad.taken {
-						elevator.AddRequest(ad.req)
-						fmt.Printf("Took my own ASSIGN message\n")
-					}
-
-				case BACKUP:
-					backup.synced = true
-
-				}
-
-			default:
-			}
-
-		} else if mode == Local {
-
-			select {
-			case req := <-reqch:
+			} else if mode == Local {
 				elevator.AddRequest(req)
-			default:
 			}
 
-		} else if mode == Broken {
+		case msg := <-msgsFromOther:
+			switch msg.Type {
+			case COST:
+				var cd costData
+				if err := unpackData(msg.Data, &cd); err != nil {
+					errorlog.Println(err)
+					break
+				}
+				
+				debug.Printf("Received cost message: \n\t%v\n", cd)
+
+				// Update cost message if our cost is lower.
+				cost := elevator.SimulateCost(cd.req)
+				if cost < cd.cost {
+					cd.elevator = node.Addr()
+					cd.cost = cost
+				}
+
+				debug.Printf("Forwarded cost message: \n\t%v\n", cd)
+
+				packData(msg.Data, &cd)
+
+			case ASSIGN:
+				var ad assignData
+				if err := unpackData(msg.Data, &ad); err != nil {
+					errorlog.Println(err)
+					break
+				}
+				
+				debug.Printf("Received assign message: \n\t%v\n", ad)
+
+				// Add request to elevator if the request was assign to this elevator.
+				if ad.elevator == node.Addr() {
+					elevator.AddRequest(ad.req)
+					panel.SetLamp(btnFromDir(ad.req.direction), ad.req.floor, true)
+					ad.taken = true
+				}
+				
+				debug.Printf("Forwarded assign message: \n\t%v\n", ad)
+				
+				packData(msg.Data, &ad)
+
+			case BACKUP:
+				var bd backupData
+				if err := unpackData(msg.Data, &bd); err != nil {
+					errorlog.Println(err)
+					break
+				}
+
+				debug.Printf("Forwarded backup message: \n\t%v\n", bd)
+
+				old := backup.backups[bd.elevator]
+				lightPanel(panel, &bd, old)
+				backup.update(&bd)
+			}
+			node.ForwardMessage(msg)
+
+		case msg := <-msgsFromThis:
+			switch msg.Type {
+			case COST:
+				var cd costData
+				if err := unpackData(msg.Data, &cd); err != nil {
+					elevator.AddRequest(req)
+					reqch = unassigned
+					errorlog.Println(err)
+					break
+				}
+
+				// Assign request to elevator with lowest cost value.
+				var ad = assignData{
+					elevator: cd.elevator,
+					req:      cd.req,
+				}
+				sendData(node, ASSIGN, &ad)
+				debug.Printf("Cost message returned: \n\t%v\n", cd)
+				debug.Printf("Sent assign message: \n\t%v\n", ad)
+
+			case ASSIGN:
+				var ad assignData
+				if err := unpackData(msg.Data, &ad); err != nil {
+					elevator.AddRequest(req)
+					reqch = unassigned
+					errorlog.Println(err)
+					break
+				}
+				
+				debug.Printf("Assign message returned: \n\t%v\n", ad)
+				if !ad.taken {
+					elevator.AddRequest(ad.req)
+				}
+
+				// Transaction complete. Ready to process next request.
+				reqch = unassigned
+
+			case BACKUP:
+				var bd backupData
+				unpackData(msg.Data, &bd)
+				debug.Printf("Backup message returned: \n\t%v\n", bd)
+
+				// Backup synchronized.
+				backup.synced = true
+
+			}
+
+		case <-interrupt:
+			elev.SetMotorDirection(elev.Stop)
+			os.Exit(0)
+
+		case dead := <-deadNode:
+			debug.Printf("%v has been disconnected. Try to find backup.\n", dead)
+
+			// Lookup backup for the disconnected elevator.
+			if deadbackup, found := backup.backups[dead]; found {
+				debug.Printf("Found backup of %v\n", dead)
+
+				// Restore requests
+				for floor := 0; floor < elev.NumFloors; floor++ {
+					for _, dir := range []elev.Direction{elev.Down, elev.Up} {
+						if deadbackup.requests[floor][indexOfDir(dir)] {
+							unassigned <- Request{floor, dir}
+						}
+					}
+				}
+			}
+		default:
 		}
 
 	}
-
-	// Stop the elevator with Ctrl+C.
-	//interruptc := make(chan os.Signal)
-	//signal.Notify(interruptc, os.Interrupt)
 
 	os.Exit(0)
 }
 
-func receiveMsgs(node *network.Node, msgs chan *network.Message) {
+// Update panel by comparing old and new backups.
+func lightPanel(panel *Panel, new, old *backupData) {
+	var empty backupData
+	if old == nil {
+		old = &empty
+	}
+	
+	for floor := 0; floor < elev.NumFloors; floor++ {
+		for _, dir := range []elev.Direction{elev.Down, elev.Up} {
+			if !old.requests[floor][indexOfDir(dir)] && new.requests[floor][indexOfDir(dir)] {
+				panel.SetLamp(btnFromDir(dir), floor, true)
+			} else if old.requests[floor][indexOfDir(dir)] && !new.requests[floor][indexOfDir(dir)] {
+				panel.SetLamp(btnFromDir(dir), floor, false)
+			}
+		}
+	}
+
+}
+
+// Listen for disconnected elevators.
+func getDeadNode(node *network.Node, c chan network.Addr) {
 	for {
-		msg := node.ReceiveMessage()
-		msgs <- msg
+		c <- node.GetDeadNode()
 	}
 }
 
+// Listen for new messages from other elevators.
+func receiveMsgs(node *network.Node, msgs chan *network.Message) {
+	for {
+		msgs <- node.ReceiveMessage()
+	}
+}
+
+// Listen for new message from this elevator.
 func receiveMyMsgs(node *network.Node, msgs chan *network.Message) {
 	for {
-		msg := node.ReceiveMyMessage()
-		msgs <- msg
+		msgs <- node.ReceiveMyMessage()
 	}
 }
